minor fixes
git-svn-id: http://svn.vuze.com/public/client@29118 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/com/aelitis/azureus/core/speedmanager/impl/SpeedManagerPingMapperImpl.java
@@ -1037,7 +1037,7 @@ SpeedManagerPingMapperImpl
 			total_deviation += deviation * deviation;
 		}
 		
-		int	deviation = (int)Math.sqrt( total_deviation / num );
+		int	deviation = (int)Math.sqrt( ((double)total_deviation) / num );
 		
 			// adjust if deviation within 50% of capacity
