+++ trunk/azureus2/src/com/aelitis/azureus/core/speedmanager/impl/v2/SpeedManagerAlgorithmProviderDHTPing.java
@@ -498,7 +498,7 @@ public class SpeedManagerAlgorithmProviderDHTPing
             consecutiveDownticks++;
 
             int lowerBound= metricBadResult - metricBadTolerance;
-            signal = (currMetricValue-lowerBound) / metricBadTolerance;
+            signal = ((float)(currMetricValue-lowerBound)) / metricBadTolerance;
             signal -= 1.0f;
         }
         else{
